requirements.txt
crewai
crewai-tools
python-decouple